+++ app/src/main/AndroidManifest.xml
@@ -9,15 +9,15 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".activity.MainActivity">
+        <activity android:name="com.madonasyombua.budgetingbuddydummy.ui.activity.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activity.LoginActivity" />
-        <activity android:name=".activity.SignupActivity"></activity>
+        <activity android:name="com.madonasyombua.budgetingbuddydummy.ui.activity.LoginActivity" />
+        <activity android:name="com.madonasyombua.budgetingbuddydummy.ui.activity.SignupActivity"></activity>
     </application>
 
 </manifest>
