--- app/src/main/java/com/madonasyombua/budgetingbuddydummy/adapters/TransactionLAdapter.kt
@@ -1,3 +1,0 @@
-package com.madonasyombua.budgetingbuddydummy.adapters
-
-
