--- app/src/main/java/com/madonasyombua/budgetingbuddydummy/adapters/TransactionListAdapter.kt
@@ -1,2 +1,0 @@
-package com.madonasyombua.budgetingbuddydummy.adapters
-
