--- app/src/main/java/com/madonasyombua/budgetingbuddydummy/contracts/BudgetContract.kt
@@ -1,3 +1,0 @@
-package com.madonasyombua.budgetingbuddydummy.contracts
-
-class BudgetContract
