--- app/src/main/java/com/madonasyombua/budgetingbuddydummy/data/BudgetData.kt
@@ -1,3 +1,0 @@
-package com.madonasyombua.budgetingbuddydummy.data
-
-class BudgetData
