+++ app/src/main/java/com/madonasyombua/budgetingbuddydummy/data/contracts/BudgetContract.kt
@@ -1,0 +1,3 @@
+package com.madonasyombua.budgetingbuddydummy.data.contracts
+
+class BudgetContract
