+++ app/src/main/java/com/madonasyombua/budgetingbuddydummy/data/models/BudgetData.kt
@@ -1,0 +1,3 @@
+package com.madonasyombua.budgetingbuddydummy.data.models
+
+class BudgetData
