+++ app/src/main/java/com/madonasyombua/budgetingbuddydummy/data/providers/BudgetProvider.kt
@@ -1,0 +1,3 @@
+package com.madonasyombua.budgetingbuddydummy.data.providers
+
+class BudgetProvider
