--- app/src/main/java/com/madonasyombua/budgetingbuddydummy/providers/BudgetProvider.kt
@@ -1,3 +1,0 @@
-package com.madonasyombua.budgetingbuddydummy.providers
-
-class BudgetProvider
