+++ app/src/main/java/com/madonasyombua/budgetingbuddydummy/ui/adapters/TransactionLAdapter.kt
@@ -1,0 +1,3 @@
+package com.madonasyombua.budgetingbuddydummy.ui.adapters
+
+
