+++ app/src/main/java/com/madonasyombua/budgetingbuddydummy/ui/adapters/TransactionListAdapter.kt
@@ -1,0 +1,2 @@
+package com.madonasyombua.budgetingbuddydummy.ui.adapters
+
