+++ app/src/main/res/layout/fragment_budgetcreationfragment_list.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".fragments.BudgetCreationFragment"
+    tools:context=".ui.fragments.BudgetCreationFragment"
     tools:listitem="@layout/fragment_budgetcreationfragment" />
